patientroom & InfectType room type is infecttype revise database... in table room store area id and infecttype id

+++ DAOModule/InfectTypeDao.cs
@@ -111,7 +111,7 @@ namespace WpfApplication1.DAOModule
         }
 
 
-        public bool DeleteInfectType(int scInfectTypeId)
+        public bool DeleteInfectType(Int64 scInfectTypeId)
         {
             try
             {
